Reverted 5625

+++ src/core/renderer/StelQGLGLSLShader.cpp
@@ -20,7 +20,6 @@
 #include "StelQGLGLSLShader.hpp"
 #include "StelQGL2Renderer.hpp"
 
-#include <stdint.h>
 
 int StelQGLGLSLShader::UNIFORM_SIZES[UniformType_max] =
 	{0, sizeof(float), sizeof(Vec2f), sizeof(Vec3f),
@@ -134,18 +133,18 @@ bool StelQGLGLSLShader::addFragmentShader(const QString& source)
 QGLShaderProgram* StelQGLGLSLShader::getProgramFromCache()
 {
 	// Add up pointers to used shaders to get the ID.
-	uintptr_t id = 0;
+	ulong id = 0;
 	foreach(QGLShader* shader, defaultVertexShaders)
 	{
-		id += reinterpret_cast<uintptr_t>(shader);
+		id += reinterpret_cast<ulong>(shader);
 	}
 	foreach(OptionalShader shader, namedVertexShaders)
 	{
-		id += shader.enabled ? reinterpret_cast<uintptr_t>(shader.shader) : 0;
+		id += shader.enabled ? reinterpret_cast<ulong>(shader.shader) : 0;
 	}
 	foreach(QGLShader* shader, defaultFragmentShaders)
 	{
-		id += reinterpret_cast<uintptr_t>(shader);
+		id += reinterpret_cast<ulong>(shader);
 	}
 
 	// If no such program in cache, return NULL
@@ -166,27 +165,27 @@ bool StelQGLGLSLShader::build()
 	// No matching program in cache, need to link a new program.
 	if(cached == NULL)
 	{
-		uintptr_t id = 0;
+		ulong id = 0;
 		QGLShaderProgram* newProgram = new QGLShaderProgram(renderer->getGLContext());
 
 		// Add all the shaders to the program.
 		foreach(QGLShader* shader, defaultVertexShaders)
 		{
 			if(!newProgram->addShader(shader)) {goto FAILED;}
-			id += reinterpret_cast<uintptr_t>(shader);
+			id += reinterpret_cast<ulong>(shader);
 		}
 		foreach(OptionalShader shader, namedVertexShaders)
 		{
 			if(shader.enabled)
 			{
 				if(!newProgram->addShader(shader.shader)) {goto FAILED;}
-				id += reinterpret_cast<uintptr_t>(shader.shader);
+				id += reinterpret_cast<ulong>(shader.shader);
 			}
 		}
 		foreach(QGLShader* shader, defaultFragmentShaders)
 		{
 			if(!newProgram->addShader(shader)) {goto FAILED;}
-			id += reinterpret_cast<uintptr_t>(shader);
+			id += reinterpret_cast<ulong>(shader);
 		}
 		Q_ASSERT_X(id > 0, Q_FUNC_INFO, "Trying to build() a StelQGLGLSLShader "
 		           "but no vertex or fragment shaders were added");
